Improved DetectWin() Algorithm
With updated VerifyConntectFour() method using string::find, there was no need to continue to take snapshots of 4 tokens at a time in either a row or column. Now I can take the 'whole picture' of a row or column and search for a winning match within that. Horizontal and vertical detection algorithms are now much easier to understand.